Fix detekt issues without @Suppress
- Split SwapUiModule into SwapUiModule and SwapTrackingModule
- Create AccountBalanceProvider to group balance-related use cases
- Reduce TransactionSignManager constructor parameters from 10 to 8
- Add cursor rule: never use @Suppress without approval
