+++ app/src/main/kotlin/com/algorand/android/core/transaction/AccountBalanceProvider.kt
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2022-2025 Pera Wallet, LDA
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License
+ */
+
+package com.algorand.android.core.transaction
+
+import com.algorand.wallet.account.core.domain.usecase.GetAccountMinBalance
+import com.algorand.wallet.account.info.domain.usecase.GetAccountAlgoBalance
+import com.algorand.wallet.account.info.domain.usecase.GetAccountAssetHoldingAmount
+import java.math.BigInteger
+import javax.inject.Inject
+
+class AccountBalanceProvider @Inject constructor(
+    private val getAccountAlgoBalance: GetAccountAlgoBalance,
+    private val getAccountAssetHoldingAmount: GetAccountAssetHoldingAmount,
+    private val getAccountMinBalance: GetAccountMinBalance
+) {
+
+    suspend fun getAlgoBalance(address: String): BigInteger? {
+        return getAccountAlgoBalance(address)
+    }
+
+    suspend fun getAssetHoldingAmount(address: String, assetId: Long): BigInteger? {
+        return getAccountAssetHoldingAmount(address, assetId)
+    }
+
+    suspend fun getMinBalance(address: String): BigInteger {
+        return getAccountMinBalance(address)
+    }
+}
